+++ cw-orch/src/osmosis_test_tube/core.rs
@@ -1,19 +1,21 @@
+use std::str::FromStr;
+
 use crate::contract::WasmPath;
 use crate::prelude::Uploadable;
+use cosmwasm_std::{coin, Addr, ContractInfoResponse, StdResult};
+
+use cw_orch_core::environment::{BankQuerier, BankSetter, WasmCodeQuerier};
 use cw_orch_traits::stargate::Stargate;
 
 use cosmwasm_std::{Binary, BlockInfo, Coin, Timestamp, Uint128};
 use cw_multi_test::AppResponse;
-use osmosis_test_tube::Account;
-use osmosis_test_tube::Bank;
-use osmosis_test_tube::ExecuteResponse;
-use osmosis_test_tube::Gamm;
-use osmosis_test_tube::Module;
-use osmosis_test_tube::Runner;
-use osmosis_test_tube::RunnerError;
-use osmosis_test_tube::SigningAccount;
-use osmosis_test_tube::Wasm;
-use std::str::FromStr;
+use osmosis_std::types::cosmos::bank::v1beta1::{QuerySupplyOfRequest, QuerySupplyOfResponse};
+use osmosis_std::types::cosmwasm::wasm::v1::{
+    QueryCodeRequest, QueryCodeResponse, QueryContractInfoRequest, QueryContractInfoResponse,
+};
+use osmosis_test_tube::{
+    Account, Bank, ExecuteResponse, Gamm, Module, Runner, RunnerError, SigningAccount, Wasm,
+};
 
 // This should be the way to import stuff.
 // But apparently osmosis-test-tube doesn't have the same dependencies as the test-tube package
@@ -25,7 +27,6 @@ use osmosis_test_tube::osmosis_std::{
 use osmosis_test_tube::OsmosisTestApp;
 use std::{cell::RefCell, fmt::Debug, rc::Rc};
 
-use cosmwasm_std::Addr;
 use serde::{de::DeserializeOwned, Serialize};
 
 use crate::{
@@ -51,7 +52,7 @@ pub use osmosis_test_tube;
 /// use cw_orch::osmosis_test_tube::osmosis_test_tube::Account;
 ///
 /// // Creates an app, creates a sender with an initial balance
-/// let tube: OsmosisTestTube = OsmosisTestTube::new(coins(1_000_000_000_000, "uosmo"));
+/// let mut tube: OsmosisTestTube = OsmosisTestTube::new(coins(1_000_000_000_000, "uosmo"));
 ///
 /// // create an additional account
 /// let account = tube.init_account(coins(1_000_000_000, "uatom")).unwrap();
@@ -77,27 +78,33 @@ fn map_err(e: RunnerError) -> CwOrchError {
 impl<S: StateInterface> OsmosisTestTube<S> {
     /// Creates an account and sets its balance
     pub fn init_account(
-        &self,
+        &mut self,
         amount: Vec<cosmwasm_std::Coin>,
     ) -> Result<Rc<SigningAccount>, CwOrchError> {
-        self.app
+        let account = self
+            .app
             .borrow()
             .init_account(&amount)
             .map_err(map_err)
-            .map(Rc::new)
+            .map(Rc::new)?;
+
+        Ok(account)
     }
 
     /// Creates accounts and sets their balance
     pub fn init_accounts(
-        &self,
+        &mut self,
         amount: Vec<cosmwasm_std::Coin>,
         account_n: u64,
     ) -> Result<Vec<Rc<SigningAccount>>, CwOrchError> {
-        self.app
+        let accounts: Vec<_> = self
+            .app
             .borrow()
             .init_accounts(&amount, account_n)
             .map_err(map_err)
-            .map(|s| s.into_iter().map(Rc::new).collect())
+            .map(|s| s.into_iter().map(Rc::new).collect())?;
+
+        Ok(accounts)
     }
 
     /// Sends coins a specific address
@@ -145,7 +152,9 @@ impl<S: StateInterface> OsmosisTestTube<S> {
             })
             .map_err(map_err)?
             .balance
-            .map(|c| Uint128::from_str(&c.amount).unwrap())
+            .map(to_cosmwasm_coin)
+            .transpose()?
+            .map(|c| c.amount)
             .unwrap_or(Uint128::zero());
         Ok(amount)
     }
@@ -163,11 +172,8 @@ impl<S: StateInterface> OsmosisTestTube<S> {
             .map_err(map_err)?
             .balances
             .into_iter()
-            .map(|c| Coin {
-                amount: Uint128::from_str(&c.amount).unwrap(),
-                denom: c.denom,
-            })
-            .collect();
+            .map(to_cosmwasm_coin)
+            .collect::<Result<Vec<_>, _>>()?;
         Ok(amount)
     }
 }
@@ -320,6 +326,65 @@ impl<S: StateInterface> TxHandler for OsmosisTestTube<S> {
     }
 }
 
+impl BankQuerier for OsmosisTestTube {
+    fn balance(
+        &self,
+        address: impl Into<String>,
+        denom: Option<String>,
+    ) -> Result<Vec<cosmwasm_std::Coin>, <Self as TxHandler>::Error> {
+        if let Some(denom) = denom {
+            Ok(vec![Coin {
+                amount: self.query_balance(&address.into(), &denom)?,
+                denom,
+            }])
+        } else {
+            self.query_all_balances(&address.into())
+        }
+    }
+
+    fn supply_of(
+        &self,
+        denom: impl Into<String>,
+    ) -> Result<cosmwasm_std::Coin, <Self as TxHandler>::Error> {
+        let denom: String = denom.into();
+        let supply_of_result: QuerySupplyOfResponse = self
+            .app
+            .borrow()
+            .query(
+                "/cosmos.bank.v1beta1.Query/SupplyOf",
+                &QuerySupplyOfRequest {
+                    denom: denom.clone(),
+                },
+            )
+            .map_err(map_err)?;
+
+        Ok(supply_of_result
+            .amount
+            .map(|c| {
+                // Ok::<_, StdError>(cosmwasm_std::Coin {
+                //     amount: c.amount.parse()?,
+                //     denom: c.denom,
+                // })
+                to_cosmwasm_coin(c)
+            })
+            .transpose()?
+            .unwrap_or(coin(0, &denom)))
+    }
+}
+
+impl BankSetter for OsmosisTestTube {
+    /// It's impossible to set the balance of an address directly in OsmosisTestTub
+    /// So for this implementation, we use a weird algorithm
+    fn set_balance(
+        &mut self,
+        _address: &Addr,
+        _amount: Vec<Coin>,
+    ) -> Result<(), <Self as TxHandler>::Error> {
+        // We check the current balance
+        unimplemented!();
+    }
+}
+
 impl Stargate for OsmosisTestTube {
     fn commit_any<R: prost::Message + Default>(
         &self,
@@ -336,5 +401,117 @@ impl Stargate for OsmosisTestTube {
             data: Some(Binary(tx_response.raw_data)),
             events: tx_response.events,
         })
+    }
+}
+
+impl WasmCodeQuerier for OsmosisTestTube {
+    fn contract_hash(&self, code_id: u64) -> Result<String, <Self as TxHandler>::Error> {
+        let code_info_result: QueryCodeResponse = self
+            .app
+            .borrow()
+            .query(
+                "/cosmwasm.wasm.v1.Query/Code",
+                &QueryCodeRequest { code_id },
+            )
+            .map_err(map_err)?;
+
+        Ok(hex::encode(
+            code_info_result
+                .code_info
+                .ok_or(CwOrchError::CodeIdNotInStore(code_id.to_string()))?
+                .data_hash,
+        ))
+    }
+
+    fn contract_info<T: cw_orch_core::contract::interface_traits::ContractInstance<Self>>(
+        &self,
+        contract: &T,
+    ) -> Result<cosmwasm_std::ContractInfoResponse, <Self as TxHandler>::Error> {
+        let result = self
+            .app
+            .borrow()
+            .query::<_, QueryContractInfoResponse>(
+                "/cosmwasm.wasm.v1.Query/ContractInfo",
+                &QueryContractInfoRequest {
+                    address: contract.address()?.to_string(),
+                },
+            )
+            .map_err(map_err)?
+            .contract_info
+            .ok_or(CwOrchError::AddrNotInStore(contract.address()?.to_string()))?;
+
+        let mut contract_info = ContractInfoResponse::default();
+        contract_info.code_id = result.code_id;
+        contract_info.creator = result.creator;
+        contract_info.admin = Some(result.admin);
+
+        contract_info.ibc_port = if result.ibc_port_id.is_empty() {
+            None
+        } else {
+            Some(result.ibc_port_id)
+        };
+
+        Ok(contract_info)
+    }
+}
+
+fn to_cosmwasm_coin(c: osmosis_std::types::cosmos::base::v1beta1::Coin) -> StdResult<Coin> {
+    Ok(Coin {
+        amount: Uint128::from_str(&c.amount)?,
+        denom: c.denom,
+    })
+}
+
+#[cfg(test)]
+pub mod tests {
+    use cosmwasm_std::{coins, ContractInfoResponse};
+    use cw_orch_core::environment::{BankQuerier, WasmCodeQuerier};
+
+    use osmosis_test_tube::Account;
+
+    use super::OsmosisTestTube;
+    use counter_contract::{msg::InstantiateMsg, CounterContract};
+    use cw_orch::prelude::*;
+
+    #[test]
+    fn wasm_querier_works() -> anyhow::Result<()> {
+        let app = OsmosisTestTube::new(coins(100_000_000_000_000, "uosmo"));
+
+        let contract = CounterContract::new("counter", app.clone());
+        contract.upload()?;
+        contract.instantiate(
+            &InstantiateMsg { count: 7 },
+            Some(&Addr::unchecked(app.sender.address())),
+            None,
+        )?;
+
+        assert_eq!(
+            contract.wasm().checksum()?,
+            app.contract_hash(contract.code_id()?)?
+        );
+
+        let contract_info = app.contract_info(&contract)?;
+        let mut target_contract_info = ContractInfoResponse::default();
+        target_contract_info.admin = Some(app.sender.address().to_string());
+        target_contract_info.code_id = contract.code_id()?;
+        target_contract_info.creator = app.sender.address().to_string();
+        target_contract_info.ibc_port = None;
+        assert_eq!(contract_info, target_contract_info);
+
+        Ok(())
+    }
+
+    #[test]
+    fn bank_querier_works() -> anyhow::Result<()> {
+        let denom = "urandom";
+        let init_coins = coins(45, denom);
+        let app = OsmosisTestTube::new(init_coins.clone());
+        let sender = app.sender.address();
+        assert_eq!(
+            app.balance(sender.clone(), Some(denom.to_string()))?,
+            init_coins
+        );
+        assert_eq!(app.supply_of(denom.to_string())?, init_coins[0]);
+        Ok(())
     }
 }