+++ packages/cw-orch-core/src/environment/cosmwasm_environment.rs
@@ -2,7 +2,7 @@
 
 use super::{ChainState, IndexResponse};
 use crate::{
-    contract::interface_traits::{CwOrchUpload, Uploadable},
+    contract::interface_traits::{ContractInstance, Uploadable},
     error::CwEnvError,
 };
 use cosmwasm_std::{Addr, BlockInfo, Coin, ContractInfoResponse};
@@ -10,8 +10,8 @@ use serde::{de::DeserializeOwned, Serialize};
 use std::fmt::Debug;
 
 /// Signals a supported execution environment for CosmWasm contracts
-pub trait CwEnv: TxHandler + Clone {}
-impl<T: TxHandler + Clone> CwEnv for T {}
+pub trait CwEnv: TxHandler + BankQuerier + WasmCodeQuerier + Clone {}
+impl<T: TxHandler + BankQuerier + WasmCodeQuerier + Clone> CwEnv for T {}
 
 /// Response type for actions on an environment
 pub type TxResponse<Chain> = <Chain as TxHandler>::Response;
@@ -92,12 +92,25 @@ pub trait TxHandler: ChainState + Clone {
     }
 }
 
-pub trait WasmCodeQuerier: CwEnv {
+pub trait WasmCodeQuerier: TxHandler + Clone {
     /// Returns the checksum of provided code_id
     fn contract_hash(&self, code_id: u64) -> Result<String, <Self as TxHandler>::Error>;
     /// Returns the code_info structure of the provided contract
-    fn contract_info<T: CwOrchUpload<Self>>(
+    fn contract_info<T: ContractInstance<Self>>(
         &self,
         contract: &T,
     ) -> Result<ContractInfoResponse, <Self as TxHandler>::Error>;
+}
+
+pub trait BankQuerier: TxHandler {
+    /// Query the bank balance of a given address
+    /// If denom is None, returns all balances
+    fn balance(
+        &self,
+        address: impl Into<String>,
+        denom: Option<String>,
+    ) -> Result<Vec<Coin>, <Self as TxHandler>::Error>;
+
+    /// Query total supply in the bank for a denom
+    fn supply_of(&self, denom: impl Into<String>) -> Result<Coin, <Self as TxHandler>::Error>;
 }