+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # cw-orchestrator Changelog
 
+## Unreleased
+
+- Add `MutCwEnv` for manipulating testing environments.
+- Add `BankQuerier` as trait bound to `CwEnv`.
+- Add `WasmCodeQuerier` as trait bound to `CwEnv`.
+- Changed the snapshot feature for better snapshot readability.
+
 ## 0.18.2
 
 - Added Snapshot-Testing
@@ -22,7 +29,7 @@
 - Added `Deploy` to prelude
 - Add ability to provide custom state in `Deploy::set_contracts_state`
 - Breaking change: remove the `&self` dependency for the `Deploy::deployed_state_file_path` method
-- Using `dirs` instead of `shellexpand` for getting the default cw-orch state dir. 
+- Using `dirs` instead of `shellexpand` for getting the default cw-orch state dir.
 - Exposed the state_dir location
 - Added better env variable management by @Kayanski
 - Added message to enable logging if not enabled
@@ -64,13 +71,13 @@
 - Added helper to modify the chain id of cw-multi-test (Mock)
 - Added a trait to be able to commit any transaction to chain (Protobuf any type)
 - Added min gas and average gas utilization for computing the tx fee
-- Added Install Readme 
+- Added Install Readme
 - Change the state file default location for relative paths `./` --> `~/.cw-orchestrator`
 - Added env variables for customizing experience
 
 ## v0.15.0
 
-- Add `add_balance` function on the `Mock` type. 
+- Add `add_balance` function on the `Mock` type.
 
 ## v0.10.0