+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `BankQuerier` as trait bound to `CwEnv`.
 - Add `WasmCodeQuerier` as trait bound to `CwEnv`.
 - Changed the snapshot feature for better snapshot readability.
+- Bumped MSRV to 1.72 because of dependency `cosmrs@0.15.0`
 
 ## 0.18.2